Removing trust comment
I've thought about this one for probably years, and I've decided to
remove it, but in transparency I'm leaving it in the commit history.

The statement is still true, but I don't feel there's any value by
calling this out in the document; The purpose of this doc is to allow
people to learn a bit more about me and how to work with me, not to
scare or frighten or threaten in any way.

Signed-off-by: John Kinsella <jlkinsel@gmail.com>